pyproject.toml
[tool.poetry]
name = "dkg"
version = "8.0.2"
description = "Python library for interacting with the OriginTrail Decentralized Knowledge Graph"
authors = ["Uladzislau Hubar <hubar.uladzislau@gmail.com>, Zvonimir Sculac <zvonimir.sculac@origin-trail.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.2"
web3 = "^7.3.0"
pandas = "^2.2.3"
rdflib = "^6.3.2"
eth-abi = "^5.1.0"
ot-pyld = "^2.1.1"
hexbytes = "^1.2.1"
eth-account = "^0.13.4"
eth-typing = "^5.0.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.4"
pre-commit = "^4.0.1"

[tools.setuptools]
packages = ["dkg"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
